Extract check
