+++ packages/core/agent/src/util.ts
@@ -2,11 +2,17 @@
 // Copyright 2023 DXOS.org
 //
 
-import fs from 'node:fs';
+import fs, { existsSync } from 'node:fs';
 import path from 'node:path';
 
+import { Trigger, asyncTimeout, waitForCondition } from '@dxos/async';
 import { DX_RUNTIME, getProfilePath } from '@dxos/client-protocol';
-import { getUnixSocket } from '@dxos/client/services';
+import { SystemStatus, fromAgent, getUnixSocket } from '@dxos/client/services';
+import { Context } from '@dxos/context';
+import { invariant } from '@dxos/invariant';
+
+import { CHECK_INTERVAL, DAEMON_START_TIMEOUT } from './defs';
+import { AgentWaitTimeoutError } from './errors';
 
 export const parseAddress = (sock: string) => {
   const [protocol, path] = sock.split('://');
@@ -22,4 +28,33 @@ export const lockFilePath = (profile: string): string => {
   const lockFile = getProfilePath(DX_RUNTIME, profile, 'lockfile');
   fs.mkdirSync(path.dirname(lockFile), { recursive: true });
   return lockFile;
+};
+
+export const waitForAgentToStart = async (profile: string) => {
+  // Wait for socket file to appear.
+  {
+    await waitForCondition({
+      condition: () => existsSync(parseAddress(getUnixSocket(profile)).path),
+      timeout: DAEMON_START_TIMEOUT,
+      interval: CHECK_INTERVAL,
+      error: new AgentWaitTimeoutError(),
+    });
+  }
+
+  // Check if agent is initialized.
+  {
+    const services = fromAgent({ profile });
+    await services.open(new Context());
+
+    const trigger = new Trigger();
+    const stream = services.services.SystemService!.queryStatus({});
+    stream.subscribe(({ status }) => {
+      invariant(status === SystemStatus.ACTIVE);
+      trigger.wake();
+    });
+    await asyncTimeout(trigger.wait(), DAEMON_START_TIMEOUT, new AgentWaitTimeoutError());
+
+    await stream.close();
+    await services.close(new Context());
+  }
 };
